+++ examples-staging/basic/admin/pages/posts/hooks.tsx
@@ -5,10 +5,10 @@ import { jsx, useTheme } from '@keystone-ui/core';
 // import {
 //   getItemPage,
 //   ItemPageHooksProps,
-// } from '@keystone-next/keystone/___internal-do-not-use-will-break-in-patch/admin-ui/pages/ItemPage';
+// } from '@keystone-6/core/___internal-do-not-use-will-break-in-patch/admin-ui/pages/ItemPage';
 
 import { Fragment } from 'react';
-import { useList } from '@keystone-next/keystone/admin-ui/context';
+import { useList } from '@keystone-6/core/admin-ui/context';
 
 export function ItemPageSidebar({ listKey, item }: any) {
   const { spacing } = useTheme();