+++ packages/keystone/src/lib/core/utils.ts
@@ -1,5 +1,6 @@
+import { Limit } from 'p-limit';
 import pluralize from 'pluralize';
-import { ItemRootValue, KeystoneConfig, KeystoneContext } from '../../types';
+import { BaseItem, KeystoneConfig, KeystoneContext } from '../../types';
 import { humanize } from '../utils';
 import { prismaError } from './graphql-errors';
 import { InitialisedList } from './types-for-lists';
@@ -27,30 +28,30 @@ type PrismaModel = {
     orderBy?: readonly Record<string, 'asc' | 'desc'>[];
     include?: Record<string, boolean>;
     select?: Record<string, any>;
-  }) => PrismaPromise<ItemRootValue[]>;
-  delete: (arg: { where: UniquePrismaFilter }) => PrismaPromise<ItemRootValue>;
-  deleteMany: (arg: { where: PrismaFilter }) => PrismaPromise<ItemRootValue>;
+  }) => PrismaPromise<BaseItem[]>;
+  delete: (arg: { where: UniquePrismaFilter }) => PrismaPromise<BaseItem>;
+  deleteMany: (arg: { where: PrismaFilter }) => PrismaPromise<BaseItem>;
   findUnique: (args: {
     where: UniquePrismaFilter;
     include?: Record<string, any>;
     select?: Record<string, any>;
-  }) => PrismaPromise<ItemRootValue | null>;
+  }) => PrismaPromise<BaseItem | null>;
   findFirst: (args: {
     where: PrismaFilter;
     include?: Record<string, any>;
     select?: Record<string, any>;
-  }) => PrismaPromise<ItemRootValue | null>;
+  }) => PrismaPromise<BaseItem | null>;
   create: (args: {
     data: Record<string, any>;
     include?: Record<string, any>;
     select?: Record<string, any>;
-  }) => PrismaPromise<ItemRootValue>;
+  }) => PrismaPromise<BaseItem>;
   update: (args: {
     where: UniquePrismaFilter;
     data: Record<string, any>;
     include?: Record<string, any>;
     select?: Record<string, any>;
-  }) => PrismaPromise<ItemRootValue>;
+  }) => PrismaPromise<BaseItem>;
 };
 
 export type UnwrapPromise<TPromise extends Promise<any>> = TPromise extends Promise<infer T>
@@ -169,3 +170,23 @@ const labelToClass = (str: string) => str.replace(/\s+/g, '');
 export function getDBFieldKeyForFieldOnMultiField(fieldKey: string, subField: string) {
   return `${fieldKey}_${subField}`;
 }
+
+// this whole thing exists because Prisma doesn't handle doing multiple writes on SQLite well
+// https://github.com/prisma/prisma/issues/2955
+// note this is keyed by the prisma client instance, not the context
+// because even across requests, we want to apply the limit on SQLite
+const writeLimits = new WeakMap<object, Limit>();
+
+export const setWriteLimit = (prismaClient: object, limit: Limit) => {
+  writeLimits.set(prismaClient, limit);
+};
+
+// this accepts the context instead of the prisma client because the prisma client on context is `any`
+// so by accepting the context, it'll be less likely the wrong thing will be passed.
+export const getWriteLimit = (context: KeystoneContext) => {
+  const limit = writeLimits.get(context.prisma);
+  if (limit === undefined) {
+    throw new Error('unexpected write limit not set for prisma client');
+  }
+  return limit;
+};