+++ tests/admin-ui-tests/utils.ts
@@ -76,7 +76,11 @@ export const deleteAllData: (projectDir: string) => Promise<void> = async (proje
 
     let prisma = new PrismaClient();
 
-    await Promise.all(Object.values(prisma).map((x: any) => x?.deleteMany?.({})));
+    await prisma.$transaction(
+      Object.values(prisma)
+        .filter((x: any) => x?.deleteMany)
+        .map((x: any) => x?.deleteMany?.({}))
+    );
 
     await prisma.$disconnect();
   } finally {
@@ -150,7 +154,7 @@ export async function generalStartKeystone(projectDir: string, command: 'start' 
     throw new Error(`No such file or directory ${projectDir}`);
   }
 
-  let keystoneProcess = execa('yarn', ['keystone-next', command], {
+  let keystoneProcess = execa('yarn', ['keystone', command], {
     cwd: projectDir,
     env: process.env,
   });