+++ docs/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @keystone-next/website
 
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`3c7a581c1`](https://github.com/keystonejs/keystone/commit/3c7a581c1e53ae49c9f74509de3927ebf2703bde), [`f4554980f`](https://github.com/keystonejs/keystone/commit/f4554980f6243a6545eee6c887d946ff25cd90e3)]:
+  - @keystone-6/fields-document@1.0.0
+
+## 4.0.0
+
+### Major Changes
+
+- [#6957](https://github.com/keystonejs/keystone/pull/6957) [`de8cf44e7`](https://github.com/keystonejs/keystone/commit/de8cf44e7b328ab98e1466d7191d9ee65a57b02a) Thanks [@bladey](https://github.com/bladey)! - Update Node engines to support current Node LTS versions, currently versions 14 and 16.
+
+### Patch Changes
+
+- Updated dependencies [[`de8cf44e7`](https://github.com/keystonejs/keystone/commit/de8cf44e7b328ab98e1466d7191d9ee65a57b02a), [`2be0e18a1`](https://github.com/keystonejs/keystone/commit/2be0e18a142c85072df1f07af11dae49e4eb8324)]:
+  - @keystone-next/fields-document@14.0.0
+
 ## 3.1.9
 
 ### Patch Changes